patch start.sh

+++ Config/epiconf/start.sh
@@ -3,8 +3,6 @@
 file=~/afs/.confs/epiconf/config.ini
 file2=~/afs/.confs/epiconf/.config
 
-cp "$file" "$file2"
-
 template=~/afs/.confs/config/i3/config.template
 preconfig=~/afs/.confs/config/i3/preconfig
 config=~/afs/.confs/config/i3/config
@@ -66,8 +64,11 @@ while IFS= read -r line || [ -n "$line" ]; do
             $polybar_right="$polybar_right $var"
         fi
 
-        sed "s|__$key\__|$value|g" "$preconfig" > "$preconfig.tmp"
-        mv "$preconfig.tmp" "$preconfig"
+        if [[ $(diff "$epiconf/config.ini" "$epiconf/.config") != "" || ! -f "$FLAG_FILE" ]]; then
+            sed "s|__$key\__|$value|g" "$preconfig" > "$preconfig.tmp"
+            mv "$preconfig.tmp" "$preconfig"
+        fi
+
         declare "$var=$value"
         export "$var"
     fi
@@ -77,7 +78,11 @@ export $polybar_left
 export $polybar_center
 export $polybar_right
 
-mv "$preconfig" "$config"
+if [[ $(diff "$epiconf/config.ini" "$epiconf/.config") != "" || ! -f "$FLAG_FILE" ]]; then
+    mv "$preconfig" "$config"
 
-i3-msg reload
-polybar-msg cmd restart
+    i3-msg reload
+    polybar-msg cmd restart
+fi
+
+cp "$file" "$file2"
